assignment-2/requirements.txt
flask
Flask-Cors
psycopg2
PyJWT
python-dotenv